tests: added file structure to single directory tests
