+++ test/find-e2e-tests/tests/serenity-tools/find-index.ts
@@ -1,3 +1,12 @@
-export * from './find-page-objects'
-export * from './find-tasks'
-export * from './find-questions'
+export * from './page-objects/homepage'
+export * from './page-objects/search-page'
+export * from './page-objects/search-results-page'
+export * from './page-objects/service-details-page'
+
+export * from './questions/homepage-questions'
+export * from './questions/search-results-questions'
+export * from './questions/service-details-questions'
+
+export * from './tasks/homepage-tasks'
+export * from './tasks/search-page-tasks'
+export * from './tasks/search-results-tasks'
