+++ test/find-e2e-tests/tests/serenity-tools/page-objects/homepage.ts
@@ -1,0 +1,6 @@
+import {By, PageElement} from '@serenity-js/web';
+
+export const startButton = () =>
+    PageElement
+        .located(By.css("[data-testid='start-button']"))
+        .describedAs('start Now Button');
