+++ test/find-e2e-tests/tests/serenity-tools/page-objects/search-page.ts
@@ -1,0 +1,11 @@
+import {By, PageElement} from '@serenity-js/web';
+
+export const postcodeSearchBox = () =>
+    PageElement
+        .located(By.id("postcode"))
+        .describedAs('the Postcode searchbox');
+
+export const postcodeSearchButton = () =>
+    PageElement
+        .located(By.css("[data-testid='button-search']"))
+        .describedAs('the search button');
