+++ test/find-e2e-tests/tests/serenity-tools/page-objects/search-results-page.ts
@@ -1,22 +1,5 @@
 import {By, PageElement} from '@serenity-js/web';
 
-export const startButton = () =>
-    PageElement
-        .located(By.css("[data-testid='start-button']"))
-        .describedAs('start Now Button');
-
-export const postcodeSearchBox = () =>
-    PageElement
-        .located(By.id("postcode"))
-        .describedAs('the Postcode searchbox');
-
-
-export const postcodeSearchButton = () =>
-    PageElement
-        .located(By.css("[data-testid='button-search']"))
-        .describedAs('the search button');
-
-
 export const laServiceInformation = () =>
     PageElement
         .located(By.css("[data-testid='service-area-[e2e]testlaserviceone']"))       
@@ -39,10 +22,4 @@ export const vcfsServiceLink = () =>
     PageElement
         .located(By.css("[data-testid='[e2e]testvcfsserviceone']"))
         .describedAs('the Test VCFS Service One link');
-
-
-export const serviceDetailsPage = () =>
-    PageElement
-        .located(By.id("main-content"))
-        .describedAs('Service Details Page');
