+++ test/find-e2e-tests/tests/serenity-tools/page-objects/service-details-page.ts
@@ -1,0 +1,7 @@
+import {By, PageElement} from '@serenity-js/web';
+
+export const serviceDetailsPage = () =>
+    PageElement
+        .located(By.id("main-content"))
+        .describedAs('Service Details Page');
+
