+++ test/find-e2e-tests/tests/serenity-tools/questions/homepage-questions.ts
@@ -1,0 +1,5 @@
+import {Page} from '@serenity-js/web';
+import {Ensure, includes} from '@serenity-js/assertions';
+
+export const isTheFindPageDisplayed = () =>
+    Ensure.that(Page.current().url().toString(), includes(process.env.BASE_URL));
