+++ test/find-e2e-tests/tests/serenity-tools/questions/service-details-questions.ts
@@ -1,0 +1,10 @@
+import {Text} from '@serenity-js/web';
+import {Ensure, includes} from '@serenity-js/assertions';
+import {Answerable} from "@serenity-js/core";
+import {serviceDetailsPage} from '../find-index';
+
+export const doesTheServiceDetailsPageContentContain = (categoryName: Answerable<string>) =>
+    Ensure.that(
+        Text.of(serviceDetailsPage()),
+        includes(categoryName)
+    );
