+++ test/find-e2e-tests/tests/serenity-tools/tasks/search-results-tasks.ts
@@ -1,0 +1,18 @@
+import {Task} from '@serenity-js/core';
+import {Click} from '@serenity-js/web';
+import {
+    laServiceLink,
+    vcfsServiceLink
+} from '../find-index';
+
+export const clickOnTheLaService = (): Task =>
+    Task.where(
+        `#actor clicks on the LA service`,
+        Click.on(laServiceLink()),
+    );
+
+export const clickOnTheVcfsService = (): Task =>
+    Task.where(
+        `#actor clicks on the LA service`,
+        Click.on(vcfsServiceLink()),
+    );
